datajson: move code to separate files
This patch separates the datajson from dataset. File like
datasets.c file were really too long after adding the datajson
feature.

Doing that, some function names implementing datajson feature have
been renamed with a Datajson prefix to simplify them.
