+++ src/datajson-md5.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 2017-2024 Open Information Security Foundation
+/* Copyright (C) 2024 Open Information Security Foundation
  *
  * You can copy, redistribute or modify this Program under the terms of
  * the GNU General Public License version 2 as published by the Free
@@ -18,43 +18,47 @@
 /**
  * \file
  *
- * \author Victor Julien <victor@inliniac.net>
+ * \author Eric Leblond <el@stamus-networks.com>
  */
 
 #include "suricata-common.h"
 #include "conf.h"
 #include "datasets.h"
 #include "datajson.h"
-#include "datasets-md5.h"
+#include "datajson-md5.h"
 #include "util-hash-lookup3.h"
 
 #include "util-thash.h"
 #include "util-print.h"
 
-int Md5StrSet(void *dst, void *src)
+int Md5StrJsonSet(void *dst, void *src)
 {
-    Md5Type *src_s = src;
-    Md5Type *dst_s = dst;
+    Md5TypeJson *src_s = src;
+    Md5TypeJson *dst_s = dst;
     memcpy(dst_s->md5, src_s->md5, sizeof(dst_s->md5));
-    dst_s->rep = src_s->rep;
+    dst_s->json.value = src_s->json.value;
+    dst_s->json.len = src_s->json.len;
     return 0;
 }
 
-bool Md5StrCompare(void *a, void *b)
+bool Md5StrJsonCompare(void *a, void *b)
 {
-    const Md5Type *as = a;
-    const Md5Type *bs = b;
+    const Md5TypeJson *as = a;
+    const Md5TypeJson *bs = b;
 
     return (memcmp(as->md5, bs->md5, sizeof(as->md5)) == 0);
 }
 
-uint32_t Md5StrHash(uint32_t hash_seed, void *s)
+uint32_t Md5StrJsonHash(uint32_t hash_seed, void *s)
 {
-    const Md5Type *str = s;
+    const Md5TypeJson *str = s;
     return hashword((uint32_t *)str->md5, sizeof(str->md5) / 4, hash_seed);
 }
 
-// data stays in hash
-void Md5StrFree(void *s)
+void Md5StrJsonFree(void *s)
 {
+    const Md5TypeJson *as = s;
+    if (as->json.value) {
+        SCFree(as->json.value);
+    }
 }
