+++ src/datajson-string.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 2017-2024 Open Information Security Foundation
+/* Copyright (C) 2024 Open Information Security Foundation
  *
  * You can copy, redistribute or modify this Program under the terms of
  * the GNU General Public License version 2 as published by the Free
@@ -18,50 +18,36 @@
 /**
  * \file
  *
- * \author Victor Julien <victor@inliniac.net>
+ * \author Eric Leblond <el@stamus-networks.com>
  */
 
 #include "suricata-common.h"
 #include "conf.h"
-#include "datasets.h"
-#include "datasets-string.h"
+#include "datajson.h"
+#include "datajson-string.h"
 #include "util-thash.h"
 #include "util-print.h"
 #include "util-hash-lookup3.h"
 #include "rust.h"
 
-#if 0
-static int StringAsAscii(const void *s, char *out, size_t out_size)
+int StringJsonAsBase64(const void *s, char *out, size_t out_size)
 {
-    const StringType *str = s;
-    uint32_t offset = 0;
-    PrintRawUriBuf(out, &offset, out_size, str->ptr, str->len);
-    if (out[0] == '\0')
-        return 0;
-    strlcat(out, "\n", out_size);
-    return strlen(out);
-}
-#endif
-
-int StringAsBase64(const void *s, char *out, size_t out_size)
-{
-    const StringType *str = s;
+    const StringTypeJson *str = s;
 
     unsigned long len = Base64EncodeBufferSize(str->len);
     uint8_t encoded_data[len];
-    if (Base64Encode((unsigned char *)str->ptr, str->len,
-        encoded_data, &len) != SC_BASE64_OK)
+    if (Base64Encode((unsigned char *)str->ptr, str->len, encoded_data, &len) != SC_BASE64_OK)
         return 0;
 
     strlcpy(out, (const char *)encoded_data, out_size);
     strlcat(out, "\n", out_size);
-    return (int)strlen(out);
+    return strlen(out);
 }
 
-int StringSet(void *dst, void *src)
+int StringJsonSet(void *dst, void *src)
 {
-    StringType *src_s = src;
-    StringType *dst_s = dst;
+    StringTypeJson *src_s = src;
+    StringTypeJson *dst_s = dst;
     SCLogDebug("dst %p src %p, src_s->ptr %p src_s->len %u", dst, src, src_s->ptr, src_s->len);
 
     dst_s->len = src_s->len;
@@ -69,15 +55,17 @@ int StringSet(void *dst, void *src)
     BUG_ON(dst_s->ptr == NULL);
     memcpy(dst_s->ptr, src_s->ptr, dst_s->len);
 
-    dst_s->rep = src_s->rep;
+    dst_s->json.value = src_s->json.value;
+    dst_s->json.len = src_s->json.len;
+
     SCLogDebug("dst %p src %p, dst_s->ptr %p dst_s->len %u", dst, src, dst_s->ptr, dst_s->len);
     return 0;
 }
 
-bool StringCompare(void *a, void *b)
+bool StringJsonCompare(void *a, void *b)
 {
-    const StringType *as = a;
-    const StringType *bs = b;
+    const StringTypeJson *as = a;
+    const StringTypeJson *bs = b;
 
     if (as->len != bs->len)
         return false;
@@ -85,21 +73,18 @@ bool StringCompare(void *a, void *b)
     return (memcmp(as->ptr, bs->ptr, as->len) == 0);
 }
 
-uint32_t StringHash(uint32_t hash_seed, void *s)
+uint32_t StringJsonHash(uint32_t hash_seed, void *s)
 {
-    StringType *str = s;
+    StringTypeJson *str = s;
     return hashlittle_safe(str->ptr, str->len, hash_seed);
 }
 
-uint32_t StringGetLength(void *s)
-{
-    StringType *str = s;
-    return str->len;
-}
-
 // base data stays in hash
-void StringFree(void *s)
+void StringJsonFree(void *s)
 {
-    StringType *str = s;
+    StringTypeJson *str = s;
     SCFree(str->ptr);
+    if (str->json.value) {
+        SCFree(str->json.value);
+    }
 }
