+++ src/datasets.c
@@ -32,7 +32,7 @@
 #include "datasets-md5.h"
 #include "datasets-sha256.h"
 #include "datasets-reputation.h"
-#include "datasets-json.h"
+#include "datajson.h"
 #include "util-conf.h"
 #include "util-mem.h"
 #include "util-thash.h"
@@ -49,16 +49,12 @@ static uint32_t set_ids = 0;
 static int DatasetAddwRep(Dataset *set, const uint8_t *data, const uint32_t data_len,
         DataRepType *rep);
 
-static int DatasetAddwJson(
-        Dataset *set, const uint8_t *data, const uint32_t data_len, DataJsonType *json);
-
 static inline void DatasetUnlockData(THashData *d)
 {
     (void) THashDecrUsecnt(d);
     THashDataUnlock(d);
 }
 static bool DatasetIsStatic(const char *save, const char *load);
-static void GetDefaultMemcap(uint64_t *memcap, uint32_t *hashsize);
 
 enum DatasetTypes DatasetGetTypeFromString(const char *s)
 {
@@ -75,7 +71,23 @@ enum DatasetTypes DatasetGetTypeFromString(const char *s)
     return DATASET_TYPE_NOTSET;
 }
 
-static Dataset *DatasetAlloc(const char *name)
+void DatasetAppendSet(Dataset *set)
+{
+    set->next = sets;
+    sets = set;
+}
+
+void DatasetLock(void)
+{
+    SCMutexLock(&sets_lock);
+}
+
+void DatasetUnlock(void)
+{
+    SCMutexUnlock(&sets_lock);
+}
+
+Dataset *DatasetAlloc(const char *name)
 {
     Dataset *set = SCCalloc(1, sizeof(*set));
     if (set) {
@@ -84,7 +96,7 @@ static Dataset *DatasetAlloc(const char *name)
     return set;
 }
 
-static Dataset *DatasetSearchByName(const char *name)
+Dataset *DatasetSearchByName(const char *name)
 {
     Dataset *set = sets;
     while (set) {
@@ -94,36 +106,6 @@ static Dataset *DatasetSearchByName(const char *name)
         set = set->next;
     }
     return NULL;
-}
-
-static int HexToRaw(const uint8_t *in, size_t ins, uint8_t *out, size_t outs)
-{
-    if (ins < 2)
-        return -1;
-    if (ins % 2 != 0)
-        return -1;
-    if (outs != ins / 2)
-        return -1;
-
-    uint8_t hash[outs];
-    memset(hash, 0, outs);
-    size_t i, x;
-    for (x = 0, i = 0; i < ins; i+=2, x++) {
-        char buf[3] = { 0, 0, 0 };
-        buf[0] = in[i];
-        buf[1] = in[i+1];
-
-        long value = strtol(buf, NULL, 16);
-        if (value >= 0 && value <= 255)
-            hash[x] = (uint8_t)value;
-        else {
-            SCLogError("hash byte out of range %ld", value);
-            return -1;
-        }
-    }
-
-    memcpy(out, hash, outs);
-    return 0;
 }
 
 static int ParseRepLine(const char *in, size_t ins, DataRepType *rep_out)
@@ -167,102 +149,6 @@ static int ParseRepLine(const char *in, size_t ins, DataRepType *rep_out)
     SCLogDebug("v %"PRIu16" raw %s", v, ptrs[0]);
 
     rep_out->value = v;
-    return 0;
-}
-
-/* return true if number is a float or an integer */
-static bool IsFloat(const char *in, size_t ins)
-{
-    char *endptr;
-    float val = strtof(in, &endptr);
-    const char *end_ins = in + ins - 1;
-    if (val != 0 && (endptr == end_ins)) {
-        return true;
-    }
-    if (val == 0 && (endptr == in)) {
-        return false;
-    } else {
-        return true;
-    }
-    return false;
-}
-
-static int ParseJsonLine(const char *in, size_t ins, DataJsonType *rep_out)
-{
-    json_error_t jerror;
-    json_t *msg = json_loads(in, 0, &jerror);
-    if (msg == NULL) {
-        /* JANSSON does not see an integer, float or a string as valid JSON.
-           So we need to exclude them from failure. */
-        if (!IsFloat(in, ins) && !((in[0] == '"') && (in[ins - 1] == '"'))) {
-            SCLogWarning("dataset: Invalid json: %s: '%s'\n", jerror.text, in);
-            return -1;
-        }
-    } else {
-        json_decref(msg);
-    }
-    rep_out->len = ins;
-    rep_out->value = SCStrndup(in, ins);
-    if (rep_out->value == NULL) {
-        return -1;
-    }
-    return 0;
-}
-
-static json_t *GetArrayObject(json_t *json, char *key)
-{
-    if (!json || !key || !json_is_object(json)) {
-        return NULL;
-    }
-
-    const char *current_key = key;
-    json_t *current = json;
-    while (current_key) {
-        const char *dot = strchr(current_key, '.');
-
-        size_t key_len = dot ? (size_t)(dot - current_key) : strlen(current_key);
-        char key_buffer[key_len + 1];
-        strlcat(key_buffer, current_key, key_len + 1);
-
-        if (json_is_object(current) == false) {
-            return NULL;
-        }
-        current = json_object_get(current, key_buffer);
-        if (current == NULL) {
-            return NULL;
-        }
-        current_key = dot ? dot + 1 : NULL;
-    }
-    return current;
-}
-
-static int ParseJsonFile(const char *file, json_t **array, char *key)
-{
-    json_t *json;
-    json_error_t error;
-    /* assume we have one single JSON element in FILE */
-    json = json_load_file(file, 0, &error);
-    if (json == NULL) {
-        FatalErrorOnInit("can't load JSON, error on line %d: %s", error.line, error.text);
-        return -1;
-    }
-
-    if (key == NULL) {
-        *array = json;
-    } else {
-        *array = GetArrayObject(json, key);
-        if (*array == NULL) {
-            json_decref(json);
-            return -1;
-        }
-        json_incref(*array);
-        json_decref(json);
-    }
-    if (!json_is_array(*array)) {
-        FatalErrorOnInit("not an array");
-        json_decref(*array);
-        return -1;
-    }
     return 0;
 }
 
@@ -340,109 +226,7 @@ static int DatasetLoadIPv4(Dataset *set)
     return 0;
 }
 
-static int DatasetJsonLoadIPv4(Dataset *set, char *json_key, char *array_key)
-{
-    if (strlen(set->load) == 0)
-        return 0;
-
-    SCLogConfig("dataset: %s loading from '%s'", set->name, set->load);
-    const char *fopen_mode = "r";
-    if (strlen(set->save) > 0 && strcmp(set->save, set->load) == 0) {
-        fopen_mode = "a+";
-    }
-
-    int add_ret;
-    uint32_t cnt = 0;
-    if (json_key == NULL) {
-        FILE *fp = fopen(set->load, fopen_mode);
-        if (fp == NULL) {
-            SCLogError("fopen '%s' failed: %s", set->load, strerror(errno));
-            return -1;
-        }
-
-        char line[1024];
-        while (fgets(line, (int)sizeof(line), fp) != NULL) {
-            char *r = strchr(line, ',');
-            if (r == NULL) {
-                FatalErrorOnInit("datajson without separator %s/%s", set->name, set->load);
-                continue;
-                /* list with JSON data */
-            } else {
-                line[strlen(line) - 1] = '\0';
-                *r = '\0';
-
-                struct in_addr in;
-                if (inet_pton(AF_INET, line, &in) != 1) {
-                    FatalErrorOnInit(
-                            "datajson IPv4 parse failed %s/%s: %s", set->name, set->load, line);
-                    continue;
-                }
-
-                r++;
-
-                DataJsonType json = { .value = NULL, .len = 0 };
-                if (ParseJsonLine(r, strlen(r), &json) < 0) {
-                    FatalErrorOnInit("bad json value for dataset %s/%s", set->name, set->load);
-                    continue;
-                }
-
-                add_ret = DatasetAddwJson(set, (const uint8_t *)&in.s_addr, 4, &json);
-                if (add_ret < 0) {
-                    FatalErrorOnInit("datajson data add failed %s/%s", set->name, set->load);
-                    continue;
-                } else if (add_ret == 0) {
-                    SCFree(json.value);
-                }
-
-                cnt++;
-            }
-        }
-        fclose(fp);
-    } else {
-        json_t *json;
-
-        if (ParseJsonFile(set->load, &json, array_key) == -1)
-            return -1;
-
-        size_t index;
-        json_t *value;
-        json_array_foreach (json, index, value) {
-            json_t *key = json_object_get(value, json_key);
-
-            if (key == NULL) {
-                FatalErrorOnInit("Can't find expected key in object");
-                continue;
-            }
-
-            const char *ip_string = json_string_value(key);
-
-            struct in_addr in;
-            if (inet_pton(AF_INET, ip_string, &in) != 1) {
-                FatalErrorOnInit(
-                        "datajson IPv4 parse failed %s/%s: %s", set->name, set->load, ip_string);
-                continue;
-            }
-            DataJsonType json = { .value = NULL, .len = 0 };
-            json.value = json_dumps(value, JSON_COMPACT);
-            json.len = strlen(json.value);
-
-            add_ret = DatasetAddwJson(set, (const uint8_t *)&in.s_addr, 4, &json);
-            if (add_ret < 0) {
-                FatalErrorOnInit("datajson data add failed %s/%s", set->name, set->load);
-                continue;
-            } else if (add_ret == 0) {
-                SCFree(json.value);
-            }
-        }
-        json_decref(json);
-    }
-    THashConsolidateMemcap(set->hash);
-
-    SCLogConfig("dataset: %s loaded %u records", set->name, cnt);
-    return 0;
-}
-
-static int ParseIpv6String(Dataset *set, const char *line, struct in6_addr *in6)
+int DatasetParseIpv6String(Dataset *set, const char *line, struct in6_addr *in6)
 {
     /* Checking IPv6 case */
     char *got_colon = strchr(line, ':');
@@ -499,7 +283,7 @@ static int DatasetLoadIPv6(Dataset *set)
             SCLogDebug("line: '%s'", line);
 
             struct in6_addr in6;
-            int ret = ParseIpv6String(set, line, &in6);
+            int ret = DatasetParseIpv6String(set, line, &in6);
             if (ret < 0) {
                 FatalErrorOnInit("unable to parse IP address");
                 continue;
@@ -519,7 +303,7 @@ static int DatasetLoadIPv6(Dataset *set)
             *r = '\0';
 
             struct in6_addr in6;
-            int ret = ParseIpv6String(set, line, &in6);
+            int ret = DatasetParseIpv6String(set, line, &in6);
             if (ret < 0) {
                 FatalErrorOnInit("unable to parse IP address");
                 continue;
@@ -544,108 +328,6 @@ static int DatasetLoadIPv6(Dataset *set)
     THashConsolidateMemcap(set->hash);
 
     fclose(fp);
-    SCLogConfig("dataset: %s loaded %u records", set->name, cnt);
-    return 0;
-}
-
-static int DatasetJsonLoadIPv6(Dataset *set, char *json_key, char *array_key)
-{
-    if (strlen(set->load) == 0)
-        return 0;
-
-    SCLogConfig("dataset: %s loading from '%s'", set->name, set->load);
-    const char *fopen_mode = "r";
-    if (strlen(set->save) > 0 && strcmp(set->save, set->load) == 0) {
-        fopen_mode = "a+";
-    }
-
-    int add_ret;
-    uint32_t cnt = 0;
-    if (json_key == NULL) {
-        FILE *fp = fopen(set->load, fopen_mode);
-        if (fp == NULL) {
-            SCLogError("fopen '%s' failed: %s", set->load, strerror(errno));
-            return -1;
-        }
-
-        char line[1024];
-        while (fgets(line, (int)sizeof(line), fp) != NULL) {
-            char *r = strchr(line, ',');
-            if (r == NULL) {
-                FatalErrorOnInit("datajson without separator %s/%s", set->name, set->load);
-                /* list with JSON data */
-            } else {
-                line[strlen(line) - 1] = '\0';
-                SCLogDebug("IPv6 with JSON line: '%s'", line);
-
-                *r = '\0';
-                struct in6_addr in6;
-                int ret = ParseIpv6String(set, line, &in6);
-                if (ret < 0) {
-                    FatalErrorOnInit("unable to parse IP address");
-                    continue;
-                }
-
-                r++;
-
-                DataJsonType json = { .value = NULL, .len = 0 };
-                if (ParseJsonLine(r, strlen(r), &json) < 0) {
-                    FatalErrorOnInit("bad json value for dataset %s/%s", set->name, set->load);
-                    continue;
-                }
-
-                add_ret = DatasetAddwJson(set, (const uint8_t *)&in6.s6_addr, 16, &json);
-                if (add_ret < 0) {
-                    FatalErrorOnInit("datajson data add failed %s/%s", set->name, set->load);
-                    continue;
-                } else if (add_ret == 0) {
-                    SCFree(json.value);
-                }
-                cnt++;
-            }
-        }
-
-        fclose(fp);
-    } else {
-        json_t *json;
-        if (ParseJsonFile(set->load, &json, array_key) == -1)
-            return -1;
-
-        size_t index;
-        json_t *value;
-        json_array_foreach (json, index, value) {
-            json_t *key = json_object_get(value, json_key);
-
-            if (key == NULL) {
-                FatalErrorOnInit("Can't find expected key in object");
-                continue;
-            }
-
-            const char *ip_string = json_string_value(key);
-
-            struct in6_addr in6;
-            int ret = ParseIpv6String(set, ip_string, &in6);
-            if (ret < 0) {
-                FatalErrorOnInit("unable to parse IP address");
-                continue;
-            }
-            DataJsonType json = { .value = NULL, .len = 0 };
-            json.value = json_dumps(value, JSON_COMPACT);
-            json.len = strlen(json.value);
-
-            add_ret = DatasetAddwJson(set, (const uint8_t *)&in6.s6_addr, 16, &json);
-            if (add_ret < 0) {
-                FatalErrorOnInit("datajson data add failed %s/%s", set->name, set->load);
-                continue;
-            } else if (add_ret == 0) {
-                SCFree(json.value);
-            }
-        }
-        json_decref(json);
-    }
-
-    THashConsolidateMemcap(set->hash);
-
     SCLogConfig("dataset: %s loaded %u records", set->name, cnt);
     return 0;
 }
@@ -724,113 +406,6 @@ static int DatasetLoadMd5(Dataset *set)
     return 0;
 }
 
-static int DatasetJsonLoadMd5(Dataset *set, char *json_key, char *array_key)
-{
-    if (strlen(set->load) == 0)
-        return 0;
-
-    SCLogConfig("dataset: %s loading from '%s'", set->name, set->load);
-    const char *fopen_mode = "r";
-    if (strlen(set->save) > 0 && strcmp(set->save, set->load) == 0) {
-        fopen_mode = "a+";
-    }
-
-    int add_ret;
-    uint32_t cnt = 0;
-    if (json_key == NULL) {
-        FILE *fp = fopen(set->load, fopen_mode);
-        if (fp == NULL) {
-            SCLogError("fopen '%s' failed: %s", set->load, strerror(errno));
-            return -1;
-        }
-
-        char line[1024];
-        while (fgets(line, (int)sizeof(line), fp) != NULL) {
-            /* straight black/white list */
-            if (strlen(line) == 33) {
-                FatalErrorOnInit("datajson without separator %s/%s", set->name, set->load);
-                continue;
-                /* list with json data */
-            } else if (strlen(line) > 33 && line[32] == ',') {
-                line[strlen(line) - 1] = '\0';
-                SCLogDebug("MD5 with JSON line: '%s'", line);
-
-                uint8_t hash[16];
-                if (HexToRaw((const uint8_t *)line, 32, hash, sizeof(hash)) < 0) {
-                    FatalErrorOnInit("bad hash for dataset %s/%s", set->name, set->load);
-                    continue;
-                }
-
-                DataJsonType json = { .value = NULL, .len = 0 };
-                if (ParseJsonLine(line + 33, strlen(line) - 33, &json) < 0) {
-                    FatalErrorOnInit("bad json for dataset %s/%s", set->name, set->load);
-                    continue;
-                }
-
-                SCLogDebug("json v:%s", json.value);
-                add_ret = DatasetAddwJson(set, hash, 16, &json);
-                if (add_ret < 0) {
-                    FatalErrorOnInit("datajson data add failed %s/%s", set->name, set->load);
-                    continue;
-                } else if (add_ret == 0) {
-                    SCFree(json.value);
-                }
-
-                cnt++;
-
-            } else {
-                FatalErrorOnInit("MD5 bad line len %u: '%s'", (uint32_t)strlen(line), line);
-                continue;
-            }
-        }
-        fclose(fp);
-    } else {
-        json_t *json;
-        if (ParseJsonFile(set->load, &json, array_key) == -1)
-            return -1;
-
-        size_t index;
-        json_t *value;
-        json_array_foreach (json, index, value) {
-            json_t *key = json_object_get(value, json_key);
-
-            if (key == NULL) {
-                FatalErrorOnInit("Can't find expected key in object");
-                continue;
-            }
-
-            const char *hash_string = json_string_value(key);
-            if (strlen(hash_string) != 32) {
-                FatalErrorOnInit("Not correct length for a hash");
-                continue;
-            }
-
-            uint8_t hash[16];
-            if (HexToRaw((const uint8_t *)hash_string, 32, hash, sizeof(hash)) < 0) {
-                FatalErrorOnInit("bad hash for dataset %s/%s", set->name, set->load);
-                continue;
-            }
-
-            DataJsonType json = { .value = NULL, .len = 0 };
-            json.value = json_dumps(value, JSON_COMPACT);
-            json.len = strlen(json.value);
-
-            add_ret = DatasetAddwJson(set, (const uint8_t *)hash_string, 16, &json);
-            if (add_ret < 0) {
-                FatalErrorOnInit("datajson data add failed %s/%s", set->name, set->load);
-                continue;
-            } else if (add_ret == 0) {
-                SCFree(json.value);
-            }
-        }
-        json_decref(json);
-    }
-    THashConsolidateMemcap(set->hash);
-
-    SCLogConfig("dataset: %s loaded %u records", set->name, cnt);
-    return 0;
-}
-
 static int DatasetLoadSha256(Dataset *set)
 {
     if (strlen(set->load) == 0)
@@ -897,109 +472,6 @@ static int DatasetLoadSha256(Dataset *set)
     THashConsolidateMemcap(set->hash);
 
     fclose(fp);
-    SCLogConfig("dataset: %s loaded %u records", set->name, cnt);
-    return 0;
-}
-
-static int DatasetJsonLoadSha256(Dataset *set, char *json_key, char *array_key)
-{
-    if (strlen(set->load) == 0)
-        return 0;
-
-    SCLogConfig("dataset: %s loading from '%s'", set->name, set->load);
-    const char *fopen_mode = "r";
-    if (strlen(set->save) > 0 && strcmp(set->save, set->load) == 0) {
-        fopen_mode = "a+";
-    }
-
-    int add_ret;
-    uint32_t cnt = 0;
-    if (json_key == NULL) {
-        FILE *fp = fopen(set->load, fopen_mode);
-        if (fp == NULL) {
-            SCLogError("fopen '%s' failed: %s", set->load, strerror(errno));
-            return -1;
-        }
-
-        char line[1024];
-        while (fgets(line, (int)sizeof(line), fp) != NULL) {
-            /* straight black/white list */
-            if (strlen(line) == 65) {
-                FatalErrorOnInit("datajson missing separator %s/%s", set->name, set->load);
-                continue;
-            } else if (strlen(line) > 65 && line[64] == ',') {
-                line[strlen(line) - 1] = '\0';
-                SCLogDebug("SHA-256 with JSON line: '%s'", line);
-
-                uint8_t hash[32];
-                if (HexToRaw((const uint8_t *)line, 64, hash, sizeof(hash)) < 0) {
-                    FatalErrorOnInit("bad hash for dataset %s/%s", set->name, set->load);
-                    continue;
-                }
-
-                DataJsonType json = { .value = NULL, .len = 0 };
-                if (ParseJsonLine(line + 65, strlen(line) - 65, &json) < 0) {
-                    FatalErrorOnInit("bad rep for dataset %s/%s", set->name, set->load);
-                    continue;
-                }
-
-                SCLogDebug("json %s", json.value);
-
-                add_ret = DatasetAddwJson(set, hash, 32, &json);
-                if (add_ret < 0) {
-                    FatalErrorOnInit("datajson data add failed %s/%s", set->name, set->load);
-                    continue;
-                } else if (add_ret == 0) {
-                    SCFree(json.value);
-                }
-                cnt++;
-            }
-        }
-        fclose(fp);
-    } else {
-        json_t *json;
-
-        if (ParseJsonFile(set->load, &json, array_key) == -1)
-            return -1;
-
-        size_t index;
-        json_t *value;
-        json_array_foreach (json, index, value) {
-            json_t *key = json_object_get(value, json_key);
-
-            if (key == NULL) {
-                FatalErrorOnInit("Can't find expected key in object");
-                continue;
-            }
-
-            const char *hash_string = json_string_value(key);
-            if (strlen(hash_string) != 64) {
-                FatalErrorOnInit("Not correct length for a hash");
-                continue;
-            }
-
-            uint8_t hash[32];
-            if (HexToRaw((const uint8_t *)hash_string, 64, hash, sizeof(hash)) < 0) {
-                FatalErrorOnInit("bad hash for dataset %s/%s", set->name, set->load);
-                continue;
-            }
-
-            DataJsonType json = { .value = NULL, .len = 0 };
-            json.value = json_dumps(value, JSON_COMPACT);
-            json.len = strlen(json.value);
-
-            add_ret = DatasetAddwJson(set, (const uint8_t *)hash_string, 32, &json);
-            if (add_ret < 0) {
-                FatalErrorOnInit("datajson data add failed %s/%s", set->name, set->load);
-                continue;
-            } else if (add_ret == 0) {
-                SCFree(json.value);
-            }
-        }
-        json_decref(json);
-    }
-    THashConsolidateMemcap(set->hash);
-
     SCLogConfig("dataset: %s loaded %u records", set->name, cnt);
     return 0;
 }
@@ -1084,112 +556,6 @@ static int DatasetLoadString(Dataset *set)
     THashConsolidateMemcap(set->hash);
 
     fclose(fp);
-    SCLogConfig("dataset: %s loaded %u records", set->name, cnt);
-    return 0;
-}
-
-static int DatasetJsonLoadString(Dataset *set, char *json_key, char *array_key)
-{
-    if (strlen(set->load) == 0)
-        return 0;
-
-    SCLogConfig("dataset: %s loading from '%s'", set->name, set->load);
-    const char *fopen_mode = "r";
-    if (strlen(set->save) > 0 && strcmp(set->save, set->load) == 0) {
-        fopen_mode = "a+";
-    }
-
-    int add_ret;
-    uint32_t cnt = 0;
-    if (json_key == NULL) {
-        FILE *fp = fopen(set->load, fopen_mode);
-        if (fp == NULL) {
-            SCLogError("fopen '%s' failed: %s", set->load, strerror(errno));
-            return -1;
-        }
-
-        char line[1024];
-        while (fgets(line, (int)sizeof(line), fp) != NULL) {
-            if (strlen(line) <= 1)
-                continue;
-
-            char *r = strchr(line, ',');
-            if (r == NULL) {
-                FatalErrorOnInit("missing separator in datajson set %s/%s", set->name, set->load);
-                continue;
-            } else {
-                line[strlen(line) - 1] = '\0';
-                SCLogDebug("line: '%s'", line);
-
-                *r = '\0';
-
-                uint32_t decoded_size = Base64DecodeBufferSize(strlen(line));
-                uint8_t decoded[decoded_size];
-                uint32_t num_decoded = Base64Decode(
-                        (const uint8_t *)line, strlen(line), Base64ModeStrict, decoded);
-                if (num_decoded == 0) {
-                    FatalErrorOnInit("bad base64 encoding %s/%s", set->name, set->load);
-                    continue;
-                }
-
-                r++;
-                SCLogDebug("r '%s'", r);
-
-                // coverity[alloc_strlen : FALSE]
-                DataJsonType json = { .value = NULL, .len = 0 };
-                if (ParseJsonLine(r, strlen(r), &json) < 0) {
-                    FatalErrorOnInit("die: bad json");
-                    continue;
-                }
-                SCLogDebug("json %s", json.value);
-
-                add_ret = DatasetAddwJson(set, (const uint8_t *)decoded, num_decoded, &json);
-                if (add_ret < 0) {
-                    FatalErrorOnInit("datajson data add failed %s/%s", set->name, set->load);
-                    continue;
-                } else if (add_ret == 0) {
-                    SCFree(json.value);
-                }
-                cnt++;
-
-                SCLogDebug("line with json %s, %s", line, r);
-            }
-        }
-        fclose(fp);
-    } else {
-        json_t *json;
-
-        if (ParseJsonFile(set->load, &json, array_key) == -1)
-            return -1;
-
-        size_t index;
-        json_t *value;
-        json_array_foreach (json, index, value) {
-            json_t *key = json_object_get(value, json_key);
-
-            if (key == NULL) {
-                FatalErrorOnInit("Can't find expected key in object");
-                continue;
-            }
-
-            const char *val = json_string_value(key);
-
-            DataJsonType json = { .value = NULL, .len = 0 };
-            json.value = json_dumps(value, JSON_COMPACT);
-            json.len = strlen(json.value);
-
-            add_ret = DatasetAddwJson(set, (const uint8_t *)val, strlen(val), &json);
-            if (add_ret < 0) {
-                FatalErrorOnInit("datajson data add failed %s/%s", set->name, set->load);
-                continue;
-            } else if (add_ret == 0) {
-                SCFree(json.value);
-            }
-        }
-        json_decref(json);
-    }
-    THashConsolidateMemcap(set->hash);
-
     SCLogConfig("dataset: %s loaded %u records", set->name, cnt);
     return 0;
 }
@@ -1315,7 +681,7 @@ Dataset *DatasetGet(const char *name, enum DatasetTypes type, const char *save, 
     char cnf_name[128];
     snprintf(cnf_name, sizeof(cnf_name), "datasets.%s.hash", name);
 
-    GetDefaultMemcap(&default_memcap, &default_hashsize);
+    DatasetGetDefaultMemcap(&default_memcap, &default_hashsize);
     switch (type) {
         case DATASET_TYPE_MD5:
             set->hash = THashInit(cnf_name, sizeof(Md5Type), Md5StrSet, Md5StrFree, Md5StrHash,
@@ -1393,135 +759,6 @@ out_err:
     return NULL;
 }
 
-Dataset *DatasetJsonGet(const char *name, enum DatasetTypes type, const char *load, uint64_t memcap,
-        uint32_t hashsize, char *json_key_value, char *json_array_key)
-{
-    uint64_t default_memcap = 0;
-    uint32_t default_hashsize = 0;
-    if (strlen(name) > DATASET_NAME_MAX_LEN) {
-        return NULL;
-    }
-
-    SCMutexLock(&sets_lock);
-    Dataset *set = DatasetSearchByName(name);
-    if (set) {
-        if (type != DATASET_TYPE_NOTSET && set->type != type) {
-            SCLogError("dataset %s already "
-                       "exists and is of type %u",
-                    set->name, set->type);
-            SCMutexUnlock(&sets_lock);
-            return NULL;
-        }
-
-        if (load == NULL || strlen(load) == 0) {
-            // OK, rule keyword doesn't have to set state/load,
-            // even when yaml set has set it.
-        } else {
-            if ((load == NULL && strlen(set->load) > 0) ||
-                    (load != NULL && strcmp(set->load, load) != 0)) {
-                SCLogError("dataset %s load mismatch: %s != %s", set->name, set->load, load);
-                SCMutexUnlock(&sets_lock);
-                return NULL;
-            }
-        }
-
-        SCMutexUnlock(&sets_lock);
-        return set;
-    } else {
-        if (type == DATASET_TYPE_NOTSET) {
-            SCLogError("dataset %s not defined", name);
-            goto out_err;
-        }
-    }
-
-    set = DatasetAlloc(name);
-    if (set == NULL) {
-        goto out_err;
-    }
-
-    strlcpy(set->name, name, sizeof(set->name));
-    set->type = type;
-    if (load && strlen(load)) {
-        strlcpy(set->load, load, sizeof(set->load));
-        SCLogDebug("set \'%s\' loading \'%s\' from \'%s\'", set->name, load, set->load);
-    }
-
-    char cnf_name[128];
-    snprintf(cnf_name, sizeof(cnf_name), "datasets.%s.hash", name);
-
-    GetDefaultMemcap(&default_memcap, &default_hashsize);
-    switch (type) {
-        case DATASET_TYPE_MD5:
-            set->hash = THashInit(cnf_name, sizeof(Md5TypeJson), Md5StrJsonSet, Md5StrJsonFree,
-                    Md5StrJsonHash, Md5StrJsonCompare, NULL, NULL, load != NULL ? 1 : 0,
-                    memcap > 0 ? memcap : default_memcap,
-                    hashsize > 0 ? hashsize : default_hashsize);
-            if (set->hash == NULL)
-                goto out_err;
-            if (DatasetJsonLoadMd5(set, json_key_value, json_array_key) < 0)
-                goto out_err;
-            break;
-        case DATASET_TYPE_STRING:
-            set->hash = THashInit(cnf_name, sizeof(StringTypeJson), StringJsonSet, StringJsonFree,
-                    StringJsonHash, StringJsonCompare, NULL, NULL, load != NULL ? 1 : 0,
-                    memcap > 0 ? memcap : default_memcap,
-                    hashsize > 0 ? hashsize : default_hashsize);
-            if (set->hash == NULL)
-                goto out_err;
-            if (DatasetJsonLoadString(set, json_key_value, json_array_key) < 0)
-                goto out_err;
-            break;
-        case DATASET_TYPE_SHA256:
-            set->hash = THashInit(cnf_name, sizeof(Sha256TypeJson), Sha256StrJsonSet,
-                    Sha256StrJsonFree, Sha256StrJsonHash, Sha256StrJsonCompare, NULL, NULL,
-                    load != NULL ? 1 : 0, memcap > 0 ? memcap : default_memcap,
-                    hashsize > 0 ? hashsize : default_hashsize);
-            if (set->hash == NULL)
-                goto out_err;
-            if (DatasetJsonLoadSha256(set, json_key_value, json_array_key) < 0)
-                goto out_err;
-            break;
-        case DATASET_TYPE_IPV4:
-            set->hash = THashInit(cnf_name, sizeof(IPv4TypeJson), IPv4JsonSet, IPv4JsonFree,
-                    IPv4JsonHash, IPv4JsonCompare, NULL, NULL, load != NULL ? 1 : 0,
-                    memcap > 0 ? memcap : default_memcap,
-                    hashsize > 0 ? hashsize : default_hashsize);
-            if (set->hash == NULL)
-                goto out_err;
-            if (DatasetJsonLoadIPv4(set, json_key_value, json_array_key) < 0)
-                goto out_err;
-            break;
-        case DATASET_TYPE_IPV6:
-            set->hash = THashInit(cnf_name, sizeof(IPv6TypeJson), IPv6JsonSet, IPv6JsonFree,
-                    IPv6JsonHash, IPv6JsonCompare, NULL, NULL, load != NULL ? 1 : 0,
-                    memcap > 0 ? memcap : default_memcap,
-                    hashsize > 0 ? hashsize : default_hashsize);
-            if (set->hash == NULL)
-                goto out_err;
-            if (DatasetJsonLoadIPv6(set, json_key_value, json_array_key) < 0)
-                goto out_err;
-            break;
-    }
-
-    SCLogDebug(
-            "set %p/%s type %u save %s load %s", set, set->name, set->type, set->save, set->load);
-
-    set->next = sets;
-    sets = set;
-
-    SCMutexUnlock(&sets_lock);
-    return set;
-out_err:
-    if (set) {
-        if (set->hash) {
-            THashShutdown(set->hash);
-        }
-        SCFree(set);
-    }
-    SCMutexUnlock(&sets_lock);
-    return NULL;
-}
-
 static bool DatasetIsStatic(const char *save, const char *load)
 {
     /* A set is static if it does not have any dynamic properties like
@@ -1582,7 +819,7 @@ void DatasetPostReloadCleanup(void)
     SCMutexUnlock(&sets_lock);
 }
 
-static void GetDefaultMemcap(uint64_t *memcap, uint32_t *hashsize)
+void DatasetGetDefaultMemcap(uint64_t *memcap, uint32_t *hashsize)
 {
     const char *str = NULL;
     if (ConfGet("datasets.defaults.memcap", &str) == 1) {
@@ -1609,7 +846,7 @@ int DatasetsInit(void)
     ConfNode *datasets = ConfGetNode("datasets");
     uint64_t default_memcap = 0;
     uint32_t default_hashsize = 0;
-    GetDefaultMemcap(&default_memcap, &default_hashsize);
+    DatasetGetDefaultMemcap(&default_memcap, &default_hashsize);
     if (datasets != NULL) {
         int list_pos = 0;
         ConfNode *iter = NULL;
@@ -1890,28 +1127,6 @@ static DataRepResultType DatasetLookupStringwRep(Dataset *set,
     return rrep;
 }
 
-static DataJsonResultType DatasetLookupStringwJson(
-        Dataset *set, const uint8_t *data, const uint32_t data_len)
-{
-    DataJsonResultType rrep = { .found = false, .json = { .value = NULL, .len = 0 } };
-
-    if (set == NULL)
-        return rrep;
-
-    StringTypeJson lookup = {
-        .ptr = (uint8_t *)data, .len = data_len, .json.value = NULL, .json.len = 0
-    };
-    THashData *rdata = THashLookupFromHash(set->hash, &lookup);
-    if (rdata) {
-        StringTypeJson *found = rdata->data;
-        rrep.found = true;
-        rrep.json = found->json;
-        DatasetUnlockData(rdata);
-        return rrep;
-    }
-    return rrep;
-}
-
 static int DatasetLookupIPv4(Dataset *set, const uint8_t *data, const uint32_t data_len)
 {
     if (set == NULL)
@@ -1948,30 +1163,6 @@ static DataRepResultType DatasetLookupIPv4wRep(
         IPv4Type *found = rdata->data;
         rrep.found = true;
         rrep.rep = found->rep;
-        DatasetUnlockData(rdata);
-        return rrep;
-    }
-    return rrep;
-}
-
-static DataJsonResultType DatasetLookupIPv4wJson(
-        Dataset *set, const uint8_t *data, const uint32_t data_len)
-{
-    DataJsonResultType rrep = { .found = false, .json = { .value = NULL, .len = 0 } };
-
-    if (set == NULL)
-        return rrep;
-
-    if (data_len != 4)
-        return rrep;
-
-    IPv4TypeJson lookup = { .json.value = NULL, .json.len = 0 };
-    memcpy(lookup.ipv4, data, data_len);
-    THashData *rdata = THashLookupFromHash(set->hash, &lookup);
-    if (rdata) {
-        IPv4TypeJson *found = rdata->data;
-        rrep.found = true;
-        rrep.json = found->json;
         DatasetUnlockData(rdata);
         return rrep;
     }
@@ -2020,31 +1211,6 @@ static DataRepResultType DatasetLookupIPv6wRep(
     return rrep;
 }
 
-static DataJsonResultType DatasetLookupIPv6wJson(
-        Dataset *set, const uint8_t *data, const uint32_t data_len)
-{
-    DataJsonResultType rrep = { .found = false, .json = { .value = NULL, .len = 0 } };
-
-    if (set == NULL)
-        return rrep;
-
-    /* We can have IPv4 or IPV6 here dur to ip.src and ip.dst implementation */
-    if (data_len != 16 && data_len != 4)
-        return rrep;
-
-    IPv6TypeJson lookup = { .json.value = NULL, .json.len = 0 };
-    memcpy(lookup.ipv6, data, data_len);
-    THashData *rdata = THashLookupFromHash(set->hash, &lookup);
-    if (rdata) {
-        IPv6TypeJson *found = rdata->data;
-        rrep.found = true;
-        rrep.json = found->json;
-        DatasetUnlockData(rdata);
-        return rrep;
-    }
-    return rrep;
-}
-
 static int DatasetLookupMd5(Dataset *set, const uint8_t *data, const uint32_t data_len)
 {
     if (set == NULL)
@@ -2063,10 +1229,10 @@ static int DatasetLookupMd5(Dataset *set, const uint8_t *data, const uint32_t da
     return 0;
 }
 
-static DataRepResultType DatasetLookupMd5wRep(
-        Dataset *set, const uint8_t *data, const uint32_t data_len, const DataRepType *rep)
+static DataRepResultType DatasetLookupMd5wRep(Dataset *set,
+        const uint8_t *data, const uint32_t data_len, const DataRepType *rep)
 {
-    DataRepResultType rrep = { .found = false, .rep = { .value = 0 } };
+    DataRepResultType rrep = { .found = false, .rep = { .value = 0 }};
 
     if (set == NULL)
         return rrep;
@@ -2074,37 +1240,13 @@ static DataRepResultType DatasetLookupMd5wRep(
     if (data_len != 16)
         return rrep;
 
-    Md5Type lookup = { .rep.value = 0 };
+    Md5Type lookup = { .rep.value = 0};
     memcpy(lookup.md5, data, data_len);
     THashData *rdata = THashLookupFromHash(set->hash, &lookup);
     if (rdata) {
         Md5Type *found = rdata->data;
         rrep.found = true;
         rrep.rep = found->rep;
-        DatasetUnlockData(rdata);
-        return rrep;
-    }
-    return rrep;
-}
-
-static DataJsonResultType DatasetLookupMd5wJson(
-        Dataset *set, const uint8_t *data, const uint32_t data_len)
-{
-    DataJsonResultType rrep = { .found = false, .json = { .value = NULL, .len = 0 } };
-
-    if (set == NULL)
-        return rrep;
-
-    if (data_len != 16)
-        return rrep;
-
-    Md5TypeJson lookup = { .json.value = NULL, .json.len = 0 };
-    memcpy(lookup.md5, data, data_len);
-    THashData *rdata = THashLookupFromHash(set->hash, &lookup);
-    if (rdata) {
-        Md5TypeJson *found = rdata->data;
-        rrep.found = true;
-        rrep.json = found->json;
         DatasetUnlockData(rdata);
         return rrep;
     }
@@ -2147,30 +1289,6 @@ static DataRepResultType DatasetLookupSha256wRep(Dataset *set,
         Sha256Type *found = rdata->data;
         rrep.found = true;
         rrep.rep = found->rep;
-        DatasetUnlockData(rdata);
-        return rrep;
-    }
-    return rrep;
-}
-
-static DataJsonResultType DatasetLookupSha256wJson(
-        Dataset *set, const uint8_t *data, const uint32_t data_len)
-{
-    DataJsonResultType rrep = { .found = false, .json = { .value = NULL, .len = 0 } };
-
-    if (set == NULL)
-        return rrep;
-
-    if (data_len != 32)
-        return rrep;
-
-    Sha256TypeJson lookup = { .json.value = NULL, .json.len = 0 };
-    memcpy(lookup.sha256, data, data_len);
-    THashData *rdata = THashLookupFromHash(set->hash, &lookup);
-    if (rdata) {
-        Sha256TypeJson *found = rdata->data;
-        rrep.found = true;
-        rrep.json = found->json;
         DatasetUnlockData(rdata);
         return rrep;
     }
@@ -2224,29 +1342,6 @@ DataRepResultType DatasetLookupwRep(Dataset *set, const uint8_t *data, const uin
             return DatasetLookupIPv4wRep(set, data, data_len, rep);
         case DATASET_TYPE_IPV6:
             return DatasetLookupIPv6wRep(set, data, data_len, rep);
-    }
-    return rrep;
-}
-
-DataJsonResultType DatasetLookupwJson(Dataset *set, const uint8_t *data, const uint32_t data_len)
-{
-    DataJsonResultType rrep = { .found = false, .json = { .value = 0 } };
-    if (set == NULL)
-        return rrep;
-
-    switch (set->type) {
-        case DATASET_TYPE_STRING:
-            return DatasetLookupStringwJson(set, data, data_len);
-        case DATASET_TYPE_MD5:
-            return DatasetLookupMd5wJson(set, data, data_len);
-        case DATASET_TYPE_SHA256:
-            return DatasetLookupSha256wJson(set, data, data_len);
-        case DATASET_TYPE_IPV4:
-            return DatasetLookupIPv4wJson(set, data, data_len);
-        case DATASET_TYPE_IPV6:
-            return DatasetLookupIPv6wJson(set, data, data_len);
-        default:
-            break;
     }
     return rrep;
 }
@@ -2370,64 +1465,6 @@ static int DatasetAddIPv6wRep(
     return -1;
 }
 
-static int DatasetAddIPv4wJson(
-        Dataset *set, const uint8_t *data, const uint32_t data_len, const DataJsonType *json)
-{
-    if (set == NULL)
-        return -1;
-
-    if (data_len < 4)
-        return -2;
-
-    IPv4TypeJson lookup = { .json = *json };
-    memcpy(lookup.ipv4, data, 4);
-    struct THashDataGetResult res = THashGetFromHash(set->hash, &lookup);
-    if (res.data) {
-        DatasetUnlockData(res.data);
-        return res.is_new ? 1 : 0;
-    }
-    return -1;
-}
-
-static int DatasetAddIPv6wJson(
-        Dataset *set, const uint8_t *data, const uint32_t data_len, const DataJsonType *json)
-{
-    if (set == NULL)
-        return -1;
-
-    if (data_len != 16)
-        return -2;
-
-    IPv6TypeJson lookup = { .json = *json };
-    memcpy(lookup.ipv6, data, 16);
-    struct THashDataGetResult res = THashGetFromHash(set->hash, &lookup);
-    if (res.data) {
-        DatasetUnlockData(res.data);
-        return res.is_new ? 1 : 0;
-    }
-    return -1;
-}
-
-/**
- *  \retval 1 data was added to the hash
- *  \retval 0 data was not added to the hash as it is already there
- *  \retval -1 failed to add data to the hash
- */
-static int DatasetAddStringwJson(
-        Dataset *set, const uint8_t *data, const uint32_t data_len, const DataJsonType *json)
-{
-    if (set == NULL)
-        return -1;
-
-    StringTypeJson lookup = { .ptr = (uint8_t *)data, .len = data_len, .json = *json };
-    struct THashDataGetResult res = THashGetFromHash(set->hash, &lookup);
-    if (res.data) {
-        DatasetUnlockData(res.data);
-        return res.is_new ? 1 : 0;
-    }
-    return -1;
-}
-
 static int DatasetAddMd5(Dataset *set, const uint8_t *data, const uint32_t data_len)
 {
     if (set == NULL)
@@ -2465,25 +1502,6 @@ static int DatasetAddMd5wRep(
     return -1;
 }
 
-static int DatasetAddMd5wJson(
-        Dataset *set, const uint8_t *data, const uint32_t data_len, const DataJsonType *json)
-{
-    if (set == NULL)
-        return -1;
-
-    if (data_len != 16)
-        return -2;
-
-    Md5TypeJson lookup = { .json = *json };
-    memcpy(lookup.md5, data, 16);
-    struct THashDataGetResult res = THashGetFromHash(set->hash, &lookup);
-    if (res.data) {
-        DatasetUnlockData(res.data);
-        return res.is_new ? 1 : 0;
-    }
-    return -1;
-}
-
 static int DatasetAddSha256wRep(
         Dataset *set, const uint8_t *data, const uint32_t data_len, const DataRepType *rep)
 {
@@ -2494,25 +1512,6 @@ static int DatasetAddSha256wRep(
         return -2;
 
     Sha256Type lookup = { .rep = *rep };
-    memcpy(lookup.sha256, data, 32);
-    struct THashDataGetResult res = THashGetFromHash(set->hash, &lookup);
-    if (res.data) {
-        DatasetUnlockData(res.data);
-        return res.is_new ? 1 : 0;
-    }
-    return -1;
-}
-
-static int DatasetAddSha256wJson(
-        Dataset *set, const uint8_t *data, const uint32_t data_len, const DataJsonType *json)
-{
-    if (set == NULL)
-        return -1;
-
-    if (data_len != 32)
-        return -2;
-
-    Sha256TypeJson lookup = { .json = *json };
     memcpy(lookup.sha256, data, 32);
     struct THashDataGetResult res = THashGetFromHash(set->hash, &lookup);
     if (res.data) {
@@ -2581,29 +1580,6 @@ static int DatasetAddwRep(Dataset *set, const uint8_t *data, const uint32_t data
     return -1;
 }
 
-static int DatasetAddwJson(
-        Dataset *set, const uint8_t *data, const uint32_t data_len, DataJsonType *json)
-{
-    if (set == NULL)
-        return -1;
-
-    switch (set->type) {
-        case DATASET_TYPE_STRING:
-            return DatasetAddStringwJson(set, data, data_len, json);
-        case DATASET_TYPE_MD5:
-            return DatasetAddMd5wJson(set, data, data_len, json);
-        case DATASET_TYPE_SHA256:
-            return DatasetAddSha256wJson(set, data, data_len, json);
-        case DATASET_TYPE_IPV4:
-            return DatasetAddIPv4wJson(set, data, data_len, json);
-        case DATASET_TYPE_IPV6:
-            return DatasetAddIPv6wJson(set, data, data_len, json);
-        default:
-            break;
-    }
-    return -1;
-}
-
 typedef int (*DatasetOpFunc)(Dataset *set, const uint8_t *data, const uint32_t data_len);
 
 static int DatasetOpSerialized(Dataset *set, const char *string, DatasetOpFunc DatasetOpString,
@@ -2651,7 +1627,7 @@ static int DatasetOpSerialized(Dataset *set, const char *string, DatasetOpFunc D
         }
         case DATASET_TYPE_IPV6: {
             struct in6_addr in6;
-            if (ParseIpv6String(set, string, &in6) != 0) {
+            if (DatasetParseIpv6String(set, string, &in6) != 0) {
                 SCLogError("Dataset failed to import %s as IPv6", string);
                 return -2;
             }
@@ -2781,189 +1757,4 @@ int DatasetRemove(Dataset *set, const uint8_t *data, const uint32_t data_len)
             return DatasetRemoveIPv6(set, data, data_len);
     }
     return -1;
-}
-
-typedef int (*DatajsonOpFunc)(
-        Dataset *set, const uint8_t *data, const uint32_t data_len, const DataJsonType *json);
-
-static int DatajsonOpSerialized(Dataset *set, const char *string, const char *json,
-        DatajsonOpFunc DatasetOpString, DatajsonOpFunc DatasetOpMd5, DatajsonOpFunc DatasetOpSha256,
-        DatajsonOpFunc DatasetOpIPv4, DatajsonOpFunc DatasetOpIPv6)
-{
-    int ret;
-
-    if (set == NULL)
-        return -1;
-    if (strlen(string) == 0)
-        return -1;
-
-    DataJsonType jvalue = { .value = NULL, .len = 0 };
-    if (json) {
-        if (ParseJsonLine(json, strlen(json), &jvalue) < 0) {
-            SCLogNotice("bad json value for dataset %s/%s", set->name, set->load);
-            return -1;
-        }
-    }
-
-    switch (set->type) {
-        case DATASET_TYPE_STRING: {
-            uint32_t decoded_size = Base64DecodeBufferSize(strlen(string));
-            uint8_t decoded[decoded_size];
-            uint32_t num_decoded = Base64Decode(
-                    (const uint8_t *)string, strlen(string), Base64ModeStrict, decoded);
-            if (num_decoded == 0)
-                goto operror;
-            ret = DatasetOpString(set, decoded, num_decoded, &jvalue);
-            if (ret <= 0) {
-                SCFree(jvalue.value);
-            }
-            return ret;
-        }
-        case DATASET_TYPE_MD5: {
-            if (strlen(string) != 32)
-                goto operror;
-            uint8_t hash[16];
-            if (HexToRaw((const uint8_t *)string, 32, hash, sizeof(hash)) < 0)
-                goto operror;
-            ret = DatasetOpMd5(set, hash, 16, &jvalue);
-            if (ret <= 0) {
-                SCFree(jvalue.value);
-            }
-            return ret;
-        }
-        case DATASET_TYPE_SHA256: {
-            if (strlen(string) != 64)
-                goto operror;
-            uint8_t hash[32];
-            if (HexToRaw((const uint8_t *)string, 64, hash, sizeof(hash)) < 0)
-                goto operror;
-            ret = DatasetOpSha256(set, hash, 32, &jvalue);
-            if (ret <= 0) {
-                SCFree(jvalue.value);
-            }
-            return ret;
-        }
-        case DATASET_TYPE_IPV4: {
-            struct in_addr in;
-            if (inet_pton(AF_INET, string, &in) != 1)
-                goto operror;
-            ret = DatasetOpIPv4(set, (uint8_t *)&in.s_addr, 4, &jvalue);
-            if (ret <= 0) {
-                SCFree(jvalue.value);
-            }
-            return ret;
-        }
-        case DATASET_TYPE_IPV6: {
-            struct in6_addr in6;
-            if (ParseIpv6String(set, string, &in6) != 0) {
-                SCLogError("Dataset failed to import %s as IPv6", string);
-                goto operror;
-            }
-            ret = DatasetOpIPv6(set, (uint8_t *)&in6.s6_addr, 16, &jvalue);
-            if (ret <= 0) {
-                SCFree(jvalue.value);
-            }
-            return ret;
-        }
-    }
-    return -1;
-operror:
-    SCFree(jvalue.value);
-    return -2;
-}
-
-/** \brief add serialized data to json set
- *  \retval int 1 added
- *  \retval int 0 already in hash
- *  \retval int -1 API error (not added)
- *  \retval int -2 DATA error
- */
-int DatajsonAddSerialized(Dataset *set, const char *value, const char *json)
-{
-    return DatajsonOpSerialized(set, value, json, DatasetAddStringwJson, DatasetAddMd5wJson,
-            DatasetAddSha256wJson, DatasetAddIPv4wJson, DatasetAddIPv6wJson);
-}
-
-/**
- *  \retval 1 data was removed from the hash
- *  \retval 0 data not removed (busy)
- *  \retval -1 data not found
- */
-static int DatajsonRemoveString(
-        Dataset *set, const uint8_t *data, const uint32_t data_len, const DataJsonType *json)
-{
-    if (set == NULL)
-        return -1;
-
-    StringTypeJson lookup = {
-        .ptr = (uint8_t *)data, .len = data_len, .json.value = NULL, .json.len = 0
-    };
-    return THashRemoveFromHash(set->hash, &lookup);
-}
-
-static int DatajsonRemoveIPv4(
-        Dataset *set, const uint8_t *data, const uint32_t data_len, const DataJsonType *json)
-{
-    if (set == NULL)
-        return -1;
-
-    if (data_len != 4)
-        return -2;
-
-    IPv4TypeJson lookup = { .json.value = NULL, .json.len = 0 };
-    memcpy(lookup.ipv4, data, 4);
-    return THashRemoveFromHash(set->hash, &lookup);
-}
-
-static int DatajsonRemoveIPv6(
-        Dataset *set, const uint8_t *data, const uint32_t data_len, const DataJsonType *json)
-{
-    if (set == NULL)
-        return -1;
-
-    if (data_len != 16)
-        return -2;
-
-    IPv6TypeJson lookup = { .json.value = NULL, .json.len = 0 };
-    memcpy(lookup.ipv6, data, 16);
-    return THashRemoveFromHash(set->hash, &lookup);
-}
-
-static int DatajsonRemoveMd5(
-        Dataset *set, const uint8_t *data, const uint32_t data_len, const DataJsonType *json)
-{
-    if (set == NULL)
-        return -1;
-
-    if (data_len != 16)
-        return -2;
-
-    Md5TypeJson lookup = { .json.value = NULL, .json.len = 0 };
-    memcpy(lookup.md5, data, 16);
-    return THashRemoveFromHash(set->hash, &lookup);
-}
-
-static int DatajsonRemoveSha256(
-        Dataset *set, const uint8_t *data, const uint32_t data_len, const DataJsonType *json)
-{
-    if (set == NULL)
-        return -1;
-
-    if (data_len != 32)
-        return -2;
-
-    Sha256TypeJson lookup = { .json.value = NULL, .json.len = 0 };
-    memcpy(lookup.sha256, data, 32);
-    return THashRemoveFromHash(set->hash, &lookup);
-}
-
-/** \brief remove serialized data from set
- *  \retval int 1 removed
- *  \retval int 0 found but busy (not removed)
- *  \retval int -1 API error (not removed)
- *  \retval int -2 DATA error */
-int DatajsonRemoveSerialized(Dataset *set, const char *string)
-{
-    return DatajsonOpSerialized(set, string, NULL, DatajsonRemoveString, DatajsonRemoveMd5,
-            DatajsonRemoveSha256, DatajsonRemoveIPv4, DatajsonRemoveIPv6);
 }
