+++ src/datasets.h
@@ -20,7 +20,6 @@
 
 #include "util-thash.h"
 #include "datasets-reputation.h"
-#include "datasets-json.h"
 
 int DatasetsInit(void);
 void DatasetsDestroy(void);
@@ -53,6 +52,11 @@ typedef struct Dataset {
 } Dataset;
 
 enum DatasetTypes DatasetGetTypeFromString(const char *s);
+void DatasetAppendSet(Dataset *set);
+Dataset *DatasetAlloc(const char *name);
+void DatasetLock(void);
+void DatasetUnlock(void);
+Dataset *DatasetSearchByName(const char *name);
 Dataset *DatasetFind(const char *name, enum DatasetTypes type);
 Dataset *DatasetGet(const char *name, enum DatasetTypes type, const char *save, const char *load,
         uint64_t memcap, uint32_t hashsize);
@@ -61,15 +65,12 @@ int DatasetRemove(Dataset *set, const uint8_t *data, const uint32_t data_len);
 int DatasetLookup(Dataset *set, const uint8_t *data, const uint32_t data_len);
 DataRepResultType DatasetLookupwRep(Dataset *set, const uint8_t *data, const uint32_t data_len,
         const DataRepType *rep);
-DataJsonResultType DatasetLookupwJson(Dataset *set, const uint8_t *data, const uint32_t data_len);
-Dataset *DatasetJsonGet(const char *name, enum DatasetTypes type, const char *load, uint64_t memcap,
-        uint32_t hashsize, char *json_key_value, char *json_object_key);
+
+void DatasetGetDefaultMemcap(uint64_t *memcap, uint32_t *hashsize);
+int DatasetParseIpv6String(Dataset *set, const char *line, struct in6_addr *in6);
 
 int DatasetAddSerialized(Dataset *set, const char *string);
 int DatasetRemoveSerialized(Dataset *set, const char *string);
 int DatasetLookupSerialized(Dataset *set, const char *string);
-
-int DatajsonAddSerialized(Dataset *set, const char *string, const char *json);
-int DatajsonRemoveSerialized(Dataset *set, const char *string);
 
 #endif /* SURICATA_DATASETS_H */
